kmonad: update to 65b501defdd0049563752f8af8c8c57f5a1ae38b
The reason for updating to an unreleased version is that the author
(and core maintainer) of the project is chronically ill[1] and unable
to spend a lot of time on kmonad.  At the same time, there is a rather
large refactor on the way, without which a new release would not make
too much sense[2].  As such, updating to what's essentially HEAD makes
sense in this case; it is not any less stable than 0.4.1 (I've
personally been running this for months without any problems).  As a
bonus, we get GHC 9.0.2 compatibility for free and do not need to
cherry pick the relevant commits.

[1]: https://github.com/kmonad/kmonad#disclaimer
[2]: https://github.com/kmonad/kmonad/pull/372